Add search functionality to Music page
Add an AutoSuggestBox to the Music page which uses the Search service in
the Core to fetch suggestions and search results.

This also makes some major restructuring of the projects:
- Win.UI - UI app that contains the Views
- Win.Logic - Library that contains the ViewModels
- Win.Tests.Unit - Unit tests for the logic library

In the future, a `Win.Tests.UI` project may be created which tests the
UI app. All projects use the namespace of their name, prefixed with `Stoffi`.