Move theme setting to a settings service
The new settings service is backed by a platform specific storage
solution, which in UWP stores settings in the user's roaming profile.

The setting for the application theme is modified to use the new
SettingsService, which is resolved using the new IoC container from
Core.

The namespaces for the view models are also changed to properly
reflect the changed name of the projects.

Fixes: #29